markdown: move versioning stories to src/stories (peer of containers)
