+++ packages/plugins/plugin-markdown/src/stories/DocumentHistory.stories.tsx
@@ -25,7 +25,7 @@ import { Branch, Version } from '@dxos/versioning';
 import { translations } from '#translations';
 import { Markdown, MarkdownCapabilities, MarkdownEvents } from '#types';
 
-import { MarkdownPlugin } from '../../MarkdownPlugin';
+import { MarkdownPlugin } from '../MarkdownPlugin';
 
 /** Minimal plugin that contributes an empty Extensions capability for stories. */
 const MarkdownExtensionsPlugin = Plugin.define(
